backend/requirements.txt
# Web框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
fastapi-cors>=0.0.6

# 数据库
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# AI集成
google-generativeai>=0.3.2

# 数据验证
pydantic>=2.7.4

# 环境配置
python-dotenv>=1.0.0

# 安全
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# 日志
loguru>=0.7.2

# 开发和测试
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2